+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 7.3.0
+* Application & microservice restructuring
+* Dependency updates & bug fixes
+* Input OTP UI changes
+* Changed refresh token type from jwt to custom string
+
 ## 7.2.3
 * Bug fix